Restructured tests.py into a directory
