+++ story/tellmeastory/tests/test_models.py
@@ -1,0 +1,64 @@
+from django.test import TestCase
+from tellmeastory.models import User
+
+class UserModelTests(TestCase):
+    def test_valid_username(self) -> None:
+        """
+        is_valid_username() should return False if the desired
+        username contains anything other than Unicode word chars,
+        numbers, spaces, hyphens, or underscores.
+        Similarly, the username should have a length of 5 <= name <= 14.
+        """
+        invalid_name_bang: str = "Woah!"
+        invalid_name_short: str = "hey"
+        invalid_name_long: str = "0123456789TenEleven"
+        valid_name: str = "God Himself"
+
+        bang: User = User(username=invalid_name_bang)
+        short: User = User(username=invalid_name_short)
+        long: User = User(username=invalid_name_long)
+        valid: User = User(username=valid_name)
+
+        self.assertIs(bang.is_valid_username(), False)
+        self.assertIs(short.is_valid_username(), False)
+        self.assertIs(long.is_valid_username(), False)
+        self.assertIs(valid.is_valid_username(), True)
+
+        return
+
+    def test_unique_username_checking(self) -> None:
+        """
+        is_unique_username() should return False if the desired
+        username is not registered to another user account.
+        """
+        desired_dup_name: str = "Cringe"
+        desired_uniq_name: str = "NotCringe"
+
+        orig_user: User = User(username=desired_dup_name)
+        orig_user.save()
+        dup_user: User = User(username=desired_dup_name)
+        new_orig_user: User = User(username=desired_uniq_name)
+
+        self.assertIs(dup_user.is_unique_username(), False)
+        self.assertIs(new_orig_user.is_unique_username(), True)
+
+        return
+
+    def test_valid_display_name(self) -> None:
+        """
+        is_valid_display_name() should return False if the desired
+        display name does not have a length of 5 <= name <= 20.
+        """
+        invalid_too_short: str = ":)"
+        invalid_too_long: str = "Not smile because that would be too short"
+        valid: str = "God Himself"
+
+        short: User = User(display_name=invalid_too_short)
+        long: User = User(display_name=invalid_too_long)
+        good: User = User(display_name=valid)
+
+        self.assertIs(short.is_valid_display_name(), False)
+        self.assertIs(long.is_valid_display_name(), False)
+        self.assertIs(good.is_valid_display_name(), True)
+
+        return
