+++ story/tellmeastory/tests/test_registration.py
@@ -1,69 +1,7 @@
 from django.http import HttpResponse
 from django.test import TestCase
 from hashlib import sha512
-from .models import User
-
-class UserModelTests(TestCase):
-    def test_valid_username(self) -> None:
-        """
-        is_valid_username() should return False if the desired
-        username contains anyhing other than Unicode word chars,
-        numbers, spaces, hyphens, or underscores.
-        Similarly, the username should have a length of 5 <= name <= 14.
-        """
-        invalid_name_bang: str = "Woah!"
-        invalid_name_short: str = "hey"
-        invalid_name_long: str = "0123456789TenEleven"
-        valid_name: str = "God Himself"
-
-        bang: User = User(username=invalid_name_bang)
-        short: User = User(username=invalid_name_short)
-        long: User = User(username=invalid_name_long)
-        valid: User = User(username=valid_name)
-
-        self.assertIs(bang.is_valid_username(), False)
-        self.assertIs(short.is_valid_username(), False)
-        self.assertIs(long.is_valid_username(), False)
-        self.assertIs(valid.is_valid_username(), True)
-
-        return
-
-    def test_unique_username_checking(self) -> None:
-        """
-        is_unique_username() should return False if the desired
-        username is not registered to another user account.
-        """
-        desired_dup_name: str = "Cringe"
-        desired_uniq_name: str = "NotCringe"
-
-        orig_user: User = User(username=desired_dup_name)
-        orig_user.save()
-        dup_user: User = User(username=desired_dup_name)
-        new_orig_user: User = User(username=desired_uniq_name)
-
-        self.assertIs(dup_user.is_unique_username(), False)
-        self.assertIs(new_orig_user.is_unique_username(), True)
-
-        return
-
-    def test_valid_display_name(self) -> None:
-        """
-        is_valid_display_name() should return False if the desired
-        display name does not have a length of 5 <= name <= 20.
-        """
-        invalid_too_short: str = ":)"
-        invalid_too_long: str = "Not smile because that would be too short"
-        valid: str = "God Himself"
-
-        short: User = User(display_name=invalid_too_short)
-        long: User = User(display_name=invalid_too_long)
-        good: User = User(display_name=valid)
-
-        self.assertIs(short.is_valid_display_name(), False)
-        self.assertIs(long.is_valid_display_name(), False)
-        self.assertIs(good.is_valid_display_name(), True)
-
-        return
+from tellmeastory.models import User
 
 class UserRegistrationViewTests(TestCase):
     def test_blank_registration_page(self) -> None:
@@ -235,103 +173,5 @@ class UserRegistrationViewTests(TestCase):
         except: pass
 
         self.assertEqual(new_user, None)
-
-        return
-
-hash_pw = lambda pw: sha512(pw.encode("utf-8")).hexdigest()
-def insert_registered_user(username: str, password: str) -> User:
-    return User.objects.create(
-        username=username,
-        password=hash_pw(password),
-        display_name=username
-    )
-
-class UserLoginViewTests(TestCase):
-    def test_blank_login_page(self) -> None:
-        """
-        Base login page should have no error messages present.
-        """
-        res: HttpResponse = self.client.get("/story/login/")
-        self.assertEqual(res.status_code, 200)
-        self.assertNotContains(res, "No account with that username.")
-        self.assertNotContains(res, "Incorrect password.")
-
-        return
-
-    def test_login_all_fields(self) -> None:
-        """
-        Successful login should give the user some kind of session token.
-        The user should be redirected to /story/index/, though this will-
-            eventually becomes /story/account/<their_username>.
-        """
-        # insert an existing user into the db to retrieve
-        inp_name: str = "Ben Johnson"
-        inp_pass: str = "#1Aaaaa"
-        johnson: User = insert_registered_user(inp_name, inp_pass)
-
-        # refer to views.login @ line 35 for why this is terrible :(
-        res: HttpResponse = self.client.post("/story/login/", data={
-            "username": inp_name,
-            "password": inp_pass
-        })
-
-        # HTTP 302 -> Redirect (Found)
-        self.assertEqual(res.status_code, 302)
-
-        # should eventually become /story/account/<their_username>
-        self.assertEqual(res.get("location"), "/story/")
-
-        # TODO: Check for session token
-
-        return
-
-    def test_login_bad_username(self) -> None:
-        """
-        Failure to login based on username should keep the user at /story/login/.
-        An error message should be present, "No account with that username."
-        The user should have no session token.
-        """
-        # we can just not make a user account for this :)
-        inp_name: str = "Berczynski"
-        inp_pass: str = "AHHHHHH"
-
-        # refer to views.login @ line 35 for why this is terrible :(
-        res: HttpResponse = self.client.post("/story/login/", data={
-            "username": inp_name,
-            "password": inp_pass
-        })
-
-        # HTTP 200 indicates no redirect
-        self.assertEqual(res.status_code, 200)
-        self.assertEqual(res.has_header("location"), False)
-        self.assertContains(res, "No account with that username.")
-
-        # TODO: Check for no session token
-
-        return
-
-    def test_login_bad_password(self) -> None:
-        """
-        Failure to login based on password should keep the user at /story/login/.
-        An error message should be present, "Incorrect password."
-        The user should have no session token.
-        """
-        inp_name: str = "Berczynski"
-        inp_pass: str = "AHHHHHH"
-        bad_pass: str = "sjlhf"
-        mark: User = insert_registered_user(inp_name, inp_pass)
-
-        # refer to views.login @ line 35 for why this is terrible :(
-        res: HttpResponse = self.client.post("/story/login/", data={
-            "username": inp_name,
-            "password": bad_pass
-        })
-
-        # HTTP 200 indicates no redirect
-        self.assertEqual(res.status_code, 200)
-        self.assertEqual(res.has_header("location"), False)
-        self.assertContains(res, "Incorrect password.")
-
-        # TODO: Check for no session token
 
         return
